Initial simple implementation of Go select form.
Currently in the util.clj class and can choose between
channels, but has no support yet for defaults or timeouts.
For cases where the channels are not yet ready, I implemented
a looping check based on .peek with a sleep and one on .poll with
a timeout. I prefer the one with .peek and sleep as it feelds like
it minimizes latency between something between added to a queue
not first on the .poll list and getting it.